requirements.txt
django-phonenumber-field
phonenumbers
django-countries
requests